requirements.txt
numpy
scipy
pandas
scikit-posthocs
elephant
neo
Pillow
PyYAML
scikit-image
seaborn
